[all-in-compose] compose package repackaging
