+++ app/src/main/java/com/litus_animae/refitted/WorkoutCalendarViewActivity.kt
@@ -7,7 +7,6 @@ import android.util.Log
 import androidx.activity.compose.setContent
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.material.MaterialTheme
 import com.google.android.gms.auth.api.signin.GoogleSignIn
@@ -19,9 +18,8 @@ import com.google.android.gms.tasks.Task
 import com.google.firebase.auth.AuthResult
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.GoogleAuthProvider
-import com.litus_animae.refitted.compose.Theme
+import com.litus_animae.refitted.compose.util.Theme
 import com.litus_animae.refitted.compose.Top
-import com.litus_animae.refitted.models.WorkoutViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.FlowPreview
