+++ app/src/main/java/com/litus_animae/refitted/compose/exercise/input/WeightButtons.kt
@@ -12,6 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.litus_animae.refitted.compose.state.Weight
+import com.litus_animae.refitted.compose.util.Theme
 import java.text.DecimalFormat
 import kotlin.math.sign
 import kotlin.math.withSign
